Don't generate duplicate errors
Use sets to deduplicate failures, and use an async cache class to ensure that we only read and parse each file once (per analyzer).

It turns out, prior to this change we were often reading the same file many times, because our caching strategy in CachingPackageJsonReader only updated the cache _after_ the file had been fully read and parsed.
